Refactor: Responsor 클래스 생성
http response를 만들어주는 responsor클래스를 만들고 적용했습니다.
